feat: Add --due-date to show (#40)
This allows you to filter reminders in a list that are due on a specific
day, so you can do:

```
reminders show "To Do" --due-date tomorrow
```

To list all reminders on the `To Do` list that are due tomorrow. This
only filters based on whatever day you ask for, even if you ask for
something more specific. I'm not sure any more granularity than that is
useful.

This also prints the index of the reminder even if all other reminders
aren't printed, this means you could print just reminder `10` or
something. This makes it easier to make the number portable for
completing things from a previous command regardless of your filtering.

Fixes https://github.com/keith/reminders-cli/issues/28